Move to a common templated CachingPdf implementation, move the passing of the data to the CMSHistErrorPropagator to CachingAddNLL
