+++ interface/CMSHistErrorPropagator.h
@@ -13,19 +13,12 @@
 #include "HiggsAnalysis/CombinedLimit/interface/Logging.h"
 #include "HiggsAnalysis/CombinedLimit/interface/SimpleCacheSentry.h"
 #include "HiggsAnalysis/CombinedLimit/interface/CMSHistFunc.h"
-
-class CMSHistErrorPropagatorV;
-
+#include "HiggsAnalysis/CombinedLimit/interface/CMSHistV.h"
 
 class CMSHistErrorPropagator : public RooAbsReal {
 public:
   CMSHistErrorPropagator();
 
-  // Possibility to add this on a per-bin basis
-  // Different types of bin errors to handle
-  // [0] = Nothing
-  // [1] = Barlow-Beeston
-  // [2] = Mixed
   CMSHistErrorPropagator(const char* name, const char* title, RooRealVar& x,
                          RooArgList const& funcs, RooArgList const& coeffs);
 
@@ -37,9 +30,6 @@ public:
 
   virtual ~CMSHistErrorPropagator() {;}
 
-
-  // std::vector<double> getErrorMultipliers(unsigned idx);
-  // std::vector<double> const& getErrorAdditions(unsigned idx);
   void applyErrorShifts(unsigned idx, FastHisto const& nominal, FastHisto & result);
 
   Double_t evaluate() const;
@@ -56,7 +46,9 @@ public:
 
   Double_t analyticalIntegral(Int_t code, const char* rangeName = 0) const;
 
-  friend class CMSHistErrorPropagatorV;
+  void setData(RooAbsData const& data) const;
+
+  friend class CMSHistV<CMSHistErrorPropagator>;
 
  protected:
   RooRealProxy x_;
@@ -70,7 +62,7 @@ public:
 
   mutable std::vector<double> coeffvals_; //!
   mutable FastHisto valsum_; //!
-  mutable FastHisto scaledvalsum_; //!
+  mutable FastHisto cache_; //!
   mutable std::vector<double> err2sum_; //!
   mutable std::vector<double> toterr_; //!
   mutable std::vector<std::vector<double>> binmods_; //!
@@ -84,29 +76,11 @@ public:
 
 
   void initialize() const;
-  void fillSumAndErr(int eval = 0) const;
+  void updateCache(int eval = 1) const;
 
 
  private:
   ClassDef(CMSHistErrorPropagator,1)
-};
-
-class CMSHistErrorPropagatorV {
- public:
-  CMSHistErrorPropagatorV(const CMSHistErrorPropagator &,
-                              const RooAbsData& data,
-                              bool includeZeroWeights = false);
-  void fill(std::vector<Double_t>& out) const;
-
- private:
-  const CMSHistErrorPropagator& hpdf_;
-  int begin_, end_, nbins_;
-  struct Block {
-    int index, begin, end;
-    Block(int i, int begin_, int end_) : index(i), begin(begin_), end(end_) {}
-  };
-  std::vector<Block> blocks_;
-  std::vector<int> bins_;
 };
 
 #endif
